pyproject.toml
[tool.poetry]
name = "layeredimage"
version = "2024"
license = "mit"
description = "Use this module to read, and write to a number of layered image formats"
authors = ["FredHappyface"]
classifiers = [
	"Environment :: Console",
	"Development Status :: 5 - Production/Stable",
	"Intended Audience :: Developers",
	"Intended Audience :: Education",
	"Natural Language :: English",
	"Operating System :: OS Independent",
	"Programming Language :: Python :: Implementation :: CPython",
	"Topic :: Multimedia :: Graphics",
	"Topic :: Software Development :: Libraries :: Python Modules",
	"Topic :: Utilities"
]
homepage = "https://github.com/FHPythonUtils/LayeredImage"
repository = "https://github.com/FHPythonUtils/LayeredImage"
documentation = "https://github.com/FHPythonUtils/LayeredImage/blob/master/README.md"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
blendmodes = "<2026,>=2024"
pylsr = "<2026,>=2024"
pyora = "<2,>=0.3.11"
pypdn = "<2,>=1.0.6"
deprecation = "<3,>=2.1.0"
psd-tools = "<2,>=1.9.30"
pillow = "<11,>=10.2.0"
gimpformats = "<2025,>=2023.1"
loguru = "<2,>=0.7.2"

[tool.poetry.group.dev.dependencies]
imgcompare = "^2.0.1"
pytest = "^7.4.4"
handsdown = "^2.1.0"
coverage = "^7.4.0"
ruff = "^0.1.13"
blackt = "^2024"
pyright = "^1.1.347"

[tool.cercis]
line-length = 100
single-quote = false
use-tabs = true
target-version = ["py38"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py38"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D2", "COM812", "ISC001", "W191", "N", "TCH", "PLR09", "ANN101"]
fixable = ["ALL"]


[tool.ruff.lint.per-file-ignores]
"**/{tests,docs,tools}/*" = ["D", "S101", "E402"]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.tox]
legacy_tox_ini = """
[tox]
env_list =
	py311
	py310
	py39
	py38

[testenv]
deps =
	imgcompare
	pytest
commands = pytest tests
"""
